genre-edit-form
Added a (unnecessary?) div that contains a title and form for editing a genre name. This div should be hidden until an edit button is clicked from the table below. When the submit is clicked in the hidden form, a POST request will be sent including the id of the genre being edited and the new name that was entered.